Rename `Cnd` to `CndClient`
The class defined in this file is really a Cnd_Client_ and not cnd
itself.
