+++ api_tests/tests/siren_schema.ts
@@ -10,8 +10,8 @@ import "../src/schema_matcher";
 import * as sirenJsonSchema from "../siren.schema.json";
 import * as rootJsonSchema from "../root.schema.json";
 import axios from "axios";
-import { SwapEntity } from "../src/cnd/payload";
-import * as siren from "../src/cnd/siren";
+import { SwapEntity } from "../src/cnd_client/payload";
+import * as siren from "../src/cnd_client/siren";
 
 describe("Siren Schema", () => {
     it(
